refactor: controller, rest api controller 분리
