+++ src/main/java/roomescape/RoomescapeController.java
@@ -1,18 +1,10 @@
 package roomescape;
 
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
+import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class RoomescapeController {
-    private final Reservations reservations;
-
-    public RoomescapeController() {
-        this.reservations = new Reservations();
-    }
 
     @GetMapping("/admin")
     public String showAdminWelcome() {
@@ -22,29 +14,5 @@ public class RoomescapeController {
     @GetMapping("/admin/reservation")
     public String showReservations() {
         return "admin/reservation-legacy";
-    }
-
-    @GetMapping("/reservations")
-    @ResponseBody
-    public List<ReservationDto> getAllReservation() {
-        return reservations.findAll()
-                .stream()
-                .map(entity -> new ReservationDto(entity.date(), entity.name(), entity.time()))
-                .toList();
-    }
-
-    @PostMapping("/reservations")
-    public ResponseEntity<ReservationDto> createReservation(@RequestBody ReservationDto reservationDto) {
-        ReservationEntity entity = new ReservationEntity(
-                reservationDto.name(), reservationDto.date(), reservationDto.time()
-        );
-        reservations.save(entity);
-        return ResponseEntity.ok().body(reservationDto);
-    }
-
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        reservations.deleteById(id);
-        return ResponseEntity.ok().build();
     }
 }
